server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.1.0
fake-useragent==2.2.0
fastapi==0.128.7
Flask==3.0.0
frozenlist==1.7.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==4.9.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.6.4
numpy==1.26.4
openai==2.7.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.4
phonenumbers==8.13.27
pluggy==1.6.0
postgrest==2.28.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.2.5
pyroaring==1.0.3
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
realtime==2.28.0
requests==2.32.4
rich==14.3.2
rsa==4.9.1
schedule==1.2.2
selenium==4.16.0
slowapi==0.1.9
limits==3.13.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.22.0
zstandard==0.25.0
